minor title updates. complete identity implementation

+++ Tempo_Social_Music/Controllers/TempoDB.cs
@@ -37,14 +37,14 @@ namespace Tempo_Social_Music.Controllers
 
         // GET: api/TempoDB/userID
         [HttpGet("userID/{userID}")]
-        public ActionResult<TempoUser> GetUser(int userID)
+        public ActionResult<FrontEndUser> GetUser(int userID)
         {
             var getUser = _context.TempoUser.FirstOrDefaultAsync(x => x.UserPk == userID).Result;
             if (getUser is null)
             {
                 return NotFound();
             }
-            return Ok(getUser);
+            return Ok(new FrontEndUser(getUser));
         }
 
         // GET: api/TempoDB/Jams/{userPK}
@@ -74,7 +74,7 @@ namespace Tempo_Social_Music.Controllers
         [HttpGet("user")]
         public async Task<ActionResult<FrontEndUser>> GetCurrentUser()
         {
-            //returns current tempoDB user parsed as front end user object
+            //returns active tempoDB user parsed as front end user object
             FrontEndUser foundUser = await ActiveFrontEndUser();
             if (foundUser is null)
             {
@@ -88,8 +88,9 @@ namespace Tempo_Social_Music.Controllers
         #region Create
         //POST: api/TempoDB/user
         //pair progammed by AL & MD
+        //allows creation of new Mixee if active user doesn't already have a handle
         [HttpPost("user")]
-        public async Task<ActionResult<TempoUser>> CreateUser(FrontEndUser newUser)
+        public async Task<ActionResult<FrontEndUser>> CreateUser(FrontEndUser newUser)
         {
             string aspNetId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             try
@@ -108,17 +109,20 @@ namespace Tempo_Social_Music.Controllers
                 return BadRequest(e);
             }
 
-            
-            _context.TempoUser.Add(new TempoUser(newUser, aspNetId));    //add new user to database
+            TempoUser newTempoUser = new TempoUser(newUser, aspNetId); //creates TempoUser with frontenduser and aspNetID string
+            _context.TempoUser.Add(newTempoUser);    //add new user to database
             await _context.SaveChangesAsync();  //save changes to database
-            return CreatedAtAction(nameof(GetUserByName), new { username = newUser.LoginName}, newUser);  
-            //redirect to user page for new user
+            return CreatedAtAction(nameof(GetUserByName), new { username = newUser.LoginName}, new FrontEndUser(newTempoUser));  
+            //allow redirect to user page for new user
 
         }
 
         //POST: api/TempoDB/addUserFriend
         //pair progreammed by AL & MD
-        [HttpPost("addUserFriend/{userString}")] //user string is of the form $"{Username1}&{Username2}"
+        //posts a new connection to the connection table between active user and userString Mixee
+        //note: connections are symmetric, which means either party can initiate or end a connection
+        //potential for harrassment or abuse. Will eventually change symmetry
+        [HttpPost("addUserFriend/{userString}")] //user string is of the Mixee that active user wants to connect with
         public async Task<ActionResult<List<FrontEndConnection>>> AddConnection(string userString)
         {
             FrontEndUser user1 = await ActiveFrontEndUser();
@@ -149,23 +153,28 @@ namespace Tempo_Social_Music.Controllers
         // POST: api/tempoDB/Jams
         // by M
         [HttpPost("Jams")]
-        public async Task<ActionResult<Favorites>> AddJam(Favorites newFave)
+        public async Task<ActionResult<Favorites>> AddJam(FrontEndFavorite newFave)
         {
-            Favorites oldFave = PreexistingFavoriteAsync(newFave).Result;
+            //add the active user to the desired new Jam
+            newFave.UserId = ActiveFrontEndUser().Result.UserPk;
+            //check if that user already has that as a Jam
+            Favorites oldFave = PreexistingFavoriteAsync(new Favorites(newFave)).Result;
             if(oldFave != null)
             {
-                return Ok(new FrontEndFavorite(oldFave));
+                //if so, no change necessary
+                return Ok(newFave);
             }
 
             if (!ModelState.IsValid)
             {
+                //if the model state doesn't work, don't add it to the DB
                 return BadRequest();
             }
 
-            _context.Favorites.Add(newFave);
-            await _context.SaveChangesAsync();
-            return CreatedAtAction(nameof(GetFavoritesAsync), new { userPK = newFave.UserId }, new FrontEndFavorite(newFave));
-            
+            _context.Favorites.Add(new Favorites(newFave)); //if all tests pass, add to the DB
+            await _context.SaveChangesAsync();              //save changes
+            return CreatedAtAction(nameof(GetFavoritesAsync), new { userPK = newFave.UserId }, newFave);
+            //allow reroute to favorites list.
         }
 
 
@@ -174,24 +183,30 @@ namespace Tempo_Social_Music.Controllers
 
         //PUT: api/TempoDB/user
         //pair progammed by MH & MD
+        //allow alteration of certain user values
         [HttpPut("user")]
-        public async Task<ActionResult<TempoUser>> UpdateUser(TempoUser newUser)
+        public async Task<ActionResult<FrontEndUser>> UpdateUser(FrontEndUser newUser)
         {
+            //identify which Mixee is active user
             FrontEndUser activeUser = await ActiveFrontEndUser();
+            //grab the old Mixee from DB, if it exist
             TempoUser oldUser = await _context.TempoUser.FindAsync(activeUser.UserPk);
             try
             {
-                if (oldUser is null || (_context.TempoUser.Select(x => x.LoginName).Contains(newUser.LoginName) && newUser.LoginName != oldUser.LoginName))
+                //old Mixee will be null if active user has no associated Mixr acct.
+                if (oldUser is null 
+                    || (_context.TempoUser.Select(x => x.LoginName).Contains(newUser.LoginName) 
+                        && newUser.LoginName != oldUser.LoginName))
                 {
                     return BadRequest();
-                    //if required fields are empty or username already exists cannot create new user.
+                    //if no Mixee exists for a given user, or new username already exists cannot apply new values.
                 }
             }
             catch (NullReferenceException e)
             {
                 return BadRequest(e);
             }
-
+            //apply any non-null values to old user
             oldUser.LoginName = string.IsNullOrEmpty(newUser.LoginName) ? oldUser.LoginName : newUser.LoginName;
             oldUser.FirstName = string.IsNullOrEmpty(newUser.FirstName) ? oldUser.FirstName : newUser.FirstName;
             oldUser.LastName = string.IsNullOrEmpty(newUser.LastName) ? oldUser.LastName : newUser.LastName;
@@ -199,12 +214,12 @@ namespace Tempo_Social_Music.Controllers
             oldUser.ZipCode = string.IsNullOrEmpty(newUser.ZipCode) ? oldUser.ZipCode : newUser.ZipCode;
             oldUser.State = string.IsNullOrEmpty(newUser.State) ? oldUser.State : newUser.State;
             oldUser.UserBio = string.IsNullOrEmpty(newUser.UserBio) ? oldUser.UserBio : newUser.UserBio;
-
-            _context.Entry(oldUser).State = EntityState.Modified;    //track modification of Tempo User
+            
+            _context.Entry(oldUser).State = EntityState.Modified;    //track modification of old TempoUser
             _context.Update(oldUser);
             await _context.SaveChangesAsync();  //save changes to database
-            return AcceptedAtAction(nameof(GetUserByName), new { username = oldUser.LoginName }, oldUser);
-            //redirect to user page for new user
+            return AcceptedAtAction(nameof(GetUserByName), new { username = oldUser.LoginName }, new FrontEndUser(oldUser));
+            //allow redirect to user page for new user
         }
 
             #endregion
@@ -213,13 +228,16 @@ namespace Tempo_Social_Music.Controllers
             #region Delete
             //DELETE: api/tempoDB/deleteUserFriend/{userString}
             //pair programmed by M and AL
+            //deletes connection between a given user and active user
             [HttpDelete("deleteUserFriend/{userString}")]
         public async Task<ActionResult> DeleteConnection(string userString)
         {
+            //gets active user
             FrontEndUser user1 = await ActiveFrontEndUser();
             Connection connection;
             try
             {
+                //uses legacy support function to make a connection from two usernames
                 connection = ParseConnectionUserString(user1.LoginName, userString);
             }
             catch (IndexOutOfRangeException)
@@ -227,33 +245,46 @@ namespace Tempo_Social_Music.Controllers
                 return BadRequest();
             }
 
+            //sees if connection already exists
             Connection oldConnection = await PreexistingConnectionAsync(connection);
             if(oldConnection is null)
             {
+                //if it doesn't, no need to delete
                 return Ok();
             }
 
+            //remove the connection from DB
             _context.Connection.Remove(oldConnection);
             await _context.SaveChangesAsync();
+            //no content expected upon successful delete, code 204
             return NoContent();
 
         }
 
         //DELETE: api/tempoDB/Jams/
         //by M
+        //deletes a Jam
         [HttpDelete("Jams/{jam}")]
-        public async Task<ActionResult> DeleteJamAsync(int jam)
+        public async Task<ActionResult> DeleteJamAsync(int jam) //jam is the primary key for the favorites table
         {
+            //gets the jam with a given key
             Favorites oldJam = await _context.Favorites.FindAsync(jam);
             if(oldJam is null)
             {
+                // if the jam doesn't exist, no need to delete
                 return Ok();
             }
 
-            _context.Favorites.Remove(oldJam);
-            await _context.SaveChangesAsync();
-            return NoContent();
+            //tests that current Mixee is the owner of the jam
+            if (oldJam.UserId == ActiveFrontEndUser().Result.UserPk)
+            {
+                _context.Favorites.Remove(oldJam);
+                await _context.SaveChangesAsync();
+                return NoContent();
+            }
 
+            //if tests fail, return unauthorized. Active user might not be Jam owner, or other edge case.
+            return Unauthorized();
         }
 
         #endregion
@@ -261,9 +292,11 @@ namespace Tempo_Social_Music.Controllers
         private Connection ParseConnectionUserString(string user1, string user2)
         {
             //parse userString into usernames
-            List<string> users = new List<string>();
-            users.Add(user1);
-            users.Add(user2);
+            List<string> users = new List<string>
+            {
+                user1,
+                user2
+            };
 
             //test to ensure 2 users were passed
             if (users.Count != 2 || users.Intersect(_context.TempoUser.Select(x => x.LoginName)).Count() != 2)
@@ -272,11 +305,13 @@ namespace Tempo_Social_Music.Controllers
             }
 
             //build a connection from two usernames
-            Connection newConnection = new Connection();
-            newConnection.MatchValue = 1;
-            //search table of users for user with username to extract userPK
-            newConnection.User1 = _context.TempoUser.FirstAsync(x => x.LoginName == users[0]).Result.UserPk;
-            newConnection.User2 = _context.TempoUser.FirstAsync(x => x.LoginName == users[1]).Result.UserPk;
+            Connection newConnection = new Connection
+            {
+                MatchValue = 1,
+                //search table of users for user with username to extract userPK
+                User1 = _context.TempoUser.FirstAsync(x => x.LoginName == users[0]).Result.UserPk,
+                User2 = _context.TempoUser.FirstAsync(x => x.LoginName == users[1]).Result.UserPk
+            };
 
             return newConnection;
         }
@@ -295,7 +330,9 @@ namespace Tempo_Social_Music.Controllers
         {
             //see if table already contains match for these users
             Favorites oldFavorite = await _context.Favorites.FirstOrDefaultAsync(x =>
-                (x.UserId == favorite.UserId && x.SpotArtist == favorite.SpotArtist && x.SpotTrack == favorite.SpotTrack)
+                (
+                x.UserId == favorite.UserId 
+                && (x.SpotArtist == favorite.SpotArtist || x.SpotTrack == favorite.SpotTrack))
                 );
             return oldFavorite;
         }
